fix: do not fail if version range cannot be fulfilled
With the initial version of this plugin, installing optionalDependencies doesn't fail the installation process any longer, however if a package does exist, but there is no version range that satisfies the range(s) requested, it still failed the installation. With this change this plugin addresses this issue and installations do not fail any longer in such cases.